move server components to src/, DataBase/
